Move admin api auth (mostly) to owin middleware
